Conflicts:
	Examples-vc2013.sln
	Examples/3D/GeometryShader/GeometryShader-vc2013.vcxproj
	Examples/3D/Object3D/Object3D-vc2013.vcxproj
	Examples/3D/Quaternion/Quaternion-vc2013.vcxproj
	Examples/3D/VertexBuffer/VertexBuffer-vc2013.vcxproj
	Examples/Display/Alpha/Alpha-vc2013.vcxproj
	Examples/Display/Basic2D/Basic2D-vc2013.vcxproj
	Examples/Display/FullScreen/FullScreen-vc2013.vcxproj
	Examples/Display/MultiWindow/MultiWindow-vc2013.vcxproj
	Examples/Display/Particle/Particle-vc2013.vcxproj
	Examples/Display/PerlinNoise/perlin_noise-vc2013.vcxproj
	Examples/Display/SineScroll/SineScroll-vc2013.vcxproj
	Examples/Display/Thread/Thread-vc2013.vcxproj
	Examples/Display/Timing/Timing-vc2013.vcxproj
	Examples/Display_Render/Blend/blend-vc2013.vcxproj
	Examples/Display_Render/Canvas/Canvas-vc2013.vcxproj
	Examples/Display_Render/CanvasAdvanced/CanvasAdvanced-vc2013.vcxproj
	Examples/Display_Render/ColorWheel/ColorWheel-vc2013.vcxproj
	Examples/Display_Render/MapMode/mapmode-vc2013.vcxproj
	Examples/Display_Render/PixelBuffer/PixelBuffer-vc2013.vcxproj
	Examples/Display_Render/PointSprite/pointsprite-vc2013.vcxproj
	Examples/Display_Render/Stencil/stencil-vc2013.vcxproj
	Examples/Display_Shaders/Bloom/Bloom-vc2013.vcxproj
	Examples/Display_Shaders/Gaussian/Gaussian-vc2013.vcxproj
	Examples/Display_Shaders/HSVSprite/HSVSprite-vc2013.vcxproj
	Examples/Display_Shaders/NightVision/NightVision-vc2013.vcxproj
	Examples/Display_Shaders/PostProcessing/PostProcessing-vc2013.vcxproj
	Examples/Display_Shaders/ShaderEffect/ShaderEffect-vc2013.vcxproj
	Examples/Display_Shaders/Shockwave/Shockwave-vc2013.vcxproj
	Examples/Display_Text/Font/Font-vc2013.vcxproj
	Examples/Display_Text/Language/Language-vc2013.vcxproj
	Examples/Display_Text/Scroller/Scroller-vc2013.vcxproj
	Examples/Display_Text/Text/Text-vc2013.vcxproj
	Examples/Display_Text/TextFade/TextFade-vc2013.vcxproj
	Examples/Game/Pacman/Pacman-vc2013.vcxproj
	Examples/Game/PacmanEvolutions/PacmanEvolutions.vcxproj
	Examples/Game/TileMap/TileMap-vc2013.vcxproj
	Examples/Input/Input/input-vc2013.vcxproj
	Examples/Sound/Sound/sound-vc2013.vcxproj
	Examples/UI/HelloWorld/Sources/helloworld.cpp
	Examples/UI/HelloWorld/Sources/program.cpp
	Sources/GL/Platform/OSX/input_device_provider_osxkeyboard.h
	Sources/GL/Platform/OSX/input_device_provider_osxmouse.h
	Sources/GL/Platform/OSX/input_device_provider_osxmouse.mm

+++ Sources/GL/Platform/AGL/cocoa_mouse_input_device_provider.cpp
@@ -22,20 +22,20 @@ bool CocoaMouseInputDeviceProvider::get_keycode(int keycode) const
     return false;
 }
 
-int CocoaMouseInputDeviceProvider::get_x() const
+float CocoaMouseInputDeviceProvider::get_x() const
 {
     if (touches.empty())
         return 0;
     else
-        return (int)touches[0].x;
+        return touches[0].x;
 }
 
-int CocoaMouseInputDeviceProvider::get_y() const
+float CocoaMouseInputDeviceProvider::get_y() const
 {
     if (touches.empty())
         return 0;
     else
-        return (int)touches[0].y;
+        return touches[0].y;
 }
 
 }
